worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Taigi-Flow Agent Worker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=0.8.0",
    "aiohttp>=3.9.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=2.32.0",
    "taibun>=1.1.7",
    "taigi-converter",
    "piper-tts>=1.4.2",
    "livekit-plugins-silero>=1.5.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pyright>=1.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worker"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = []

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv.sources]
taigi-converter = { git = "https://github.com/Yizhe0407/HanloFlow.git" }

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-httpserver>=1.1.5",
]
